gcp/rag-scripts/requirements.txt
google-cloud-bigquery==3.27.0
google-cloud-aiplatform==1.71.1
google-cloud-storage==2.19.0
